backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
python-jose[cryptography]
python-multipart
python-dotenv
bcrypt
pydantic[email]